Refactring address-validator to use as common for all routes.
