+++ frontend/app/.server/constants/types.constant.ts
@@ -71,12 +71,8 @@ import type { ConfigFactory, LogFactory } from '~/.server/factories';
 import type { HttpClient } from '~/.server/http';
 import type { InstrumentationService } from '~/.server/observability';
 import type { BenefitApplicationStateMapper, BenefitRenewalStateMapper } from '~/.server/routes/mappers';
-import type { MailingAddressValidatorFactory } from '~/.server/routes/public/address-validation';
-import type { MailingAddressValidatorFactoryAdult } from '~/.server/routes/public/renew/adult';
-import type { MailingAddressValidatorFactoryChild } from '~/.server/routes/public/renew/child';
-import type { MailingAddressValidatorFactoryIta } from '~/.server/routes/public/renew/ita';
 import type { SecurityHandler } from '~/.server/routes/security';
-import type { AddressValidatorFactory } from '~/.server/routes/validators';
+import type { AddressValidatorFactory, MailingAddressValidatorFactory } from '~/.server/routes/validators';
 import { assignServiceIdentifiers, serviceIdentifier as serviceId } from '~/.server/utils/service-identifier.utils';
 import type { HCaptchaDtoMapper } from '~/.server/web/mappers';
 import type { DynatraceDtoMapper } from '~/.server/web/mappers/dynatrace.dto.mapper';
@@ -227,27 +223,12 @@ export const TYPES = assignServiceIdentifiers({
       BenefitApplicationStateMapper: serviceId<BenefitApplicationStateMapper>(),
       BenefitRenewalStateMapper: serviceId<BenefitRenewalStateMapper>(),
     },
-    public: {
-      addressValidation: {
-        MailingAddressValidatorFactory: serviceId<MailingAddressValidatorFactory>(),
-      },
-      renew: {
-        ita: {
-          MailingAddressValidatorFactoryIta: serviceId<MailingAddressValidatorFactoryIta>(),
-        },
-        child: {
-          MailingAddressValidatorFactoryChild: serviceId<MailingAddressValidatorFactoryChild>(),
-        },
-        adult: {
-          MailingAddressValidatorFactoryAdult: serviceId<MailingAddressValidatorFactoryAdult>(),
-        },
-      },
-    },
     security: {
       SecurityHandler: serviceId<SecurityHandler>(),
     },
     validators: {
       AddressValidatorFactory: serviceId<AddressValidatorFactory>(),
+      MailingAddressValidatorFactory: serviceId<MailingAddressValidatorFactory>(),
     },
   },
   web: {
