--- frontend/app/.server/routes/public/renew/child/index.ts
@@ -1,2 +1,0 @@
-export * from './mailing-address.validator';
-export * from './mailing-address.validator.factory';
