+++ frontend/app/.server/routes/validators/index.ts
@@ -1,2 +1,4 @@
 export * from './address.validator';
 export * from './address.validator.factory';
+export * from './mailing-address.validator';
+export * from './mailing-address.validator.factory';
