+++ frontend/app/.server/routes/validators/mailing-address.validator.factory.ts
@@ -3,35 +3,35 @@ import moize from 'moize';
 
 import { TYPES } from '~/.server/constants';
 import type { LogFactory, Logger } from '~/.server/factories';
-import { DefaultMailingAddressValidatorChild } from '~/.server/routes/public/renew/child/mailing-address.validator';
-import type { MailingAddressValidatorChild } from '~/.server/routes/public/renew/child/mailing-address.validator';
 import type { AddressValidatorFactory } from '~/.server/routes/validators/';
+import { DefaultMailingAddressValidator } from '~/.server/routes/validators/mailing-address.validator';
+import type { MailingAddressValidator } from '~/.server/routes/validators/mailing-address.validator';
 
 /**
  * Factory interface for creating mailing address validators.
  */
-export interface MailingAddressValidatorFactoryChild {
+export interface MailingAddressValidatorFactory {
   /**
    * Creates a new mailing address validator for the given locale.
    * @param locale The locale to use for validation.
    * @returns A new mailing address validator.
    */
-  createMailingAddressValidator(locale: AppLocale): MailingAddressValidatorChild;
+  createMailingAddressValidator(locale: AppLocale): MailingAddressValidator;
 }
 
 @injectable()
-export class DefaultMailingAddressValidatorFactoryChild implements MailingAddressValidatorFactoryChild {
+export class DefaultMailingAddressValidatorFactory implements MailingAddressValidatorFactory {
   private readonly log: Logger;
 
   constructor(
     @inject(TYPES.factories.LogFactory) logFactory: LogFactory,
     @inject(TYPES.routes.validators.AddressValidatorFactory) private readonly addressValidatorFactory: AddressValidatorFactory,
   ) {
-    this.log = logFactory.createLogger(this.constructor.name);
+    this.log = logFactory.createLogger('DefaultMailingAddressValidatorFactory');
     this.init();
   }
 
-  private init() {
+  private init(): void {
     this.createMailingAddressValidator = moize(this.createMailingAddressValidator, {
       maxSize: Infinity,
       onCacheAdd: (cache) => {
@@ -39,14 +39,14 @@ export class DefaultMailingAddressValidatorFactoryChild implements MailingAddres
       },
     });
 
-    this.log.debug('%s initiated.', this.constructor.name);
+    this.log.debug('DefaultMailingAddressValidatorFactory initiated.');
   }
 
   /**
    * Memoizes the creation of mailing address validators to reduce the number of times i18next files are read.
    * Each locale will have its own memoized validator.
    */
-  createMailingAddressValidator(locale: AppLocale): MailingAddressValidatorChild {
-    return new DefaultMailingAddressValidatorChild(locale, this.addressValidatorFactory);
+  createMailingAddressValidator(locale: AppLocale): MailingAddressValidator {
+    return new DefaultMailingAddressValidator(locale, this.addressValidatorFactory);
   }
 }
