+++ frontend/public/locales/en/renew-adult.json
@@ -149,17 +149,6 @@
     "cancel-btn": "Cancel",
     "save-btn": "Save",
     "error-message": {
-      "mailing-address": {
-        "address-required": "Enter mailing address, typically number and street",
-        "city-required": "Enter a city or town for your mailing address",
-        "country-required": "Select a country for your mailing address",
-        "postal-code-required": "Enter postal code or zip code for your mailing address",
-        "postal-code-valid": "Enter mailing address postal code in the correct format, such as A1A 1A1",
-        "province-required": "Select a province, territory, state or region for your mailing address",
-        "zip-code-valid": "Enter mailing address zip code in the correct format, such as 12345 or 12345-6789",
-        "invalid-postal-code-for-country": "A Canadian postal code in the mailing address was entered but Canada was not selected as the country. Please select Canada, and the proper province, or change the postal code",
-        "invalid-postal-code-for-province": "The postal code does not match the province or territory you selected"
-      },
       "home-address": {
         "address-required": "Enter home address, typically number and street",
         "city-required": "Enter a city or town for your home address",
