+++ frontend/public/locales/fr/common.json
@@ -1,0 +1,14 @@
+{
+  "error-message": {
+    "address-required": "Entrez l'adresse postale, normalement le numéro et le nom de la rue",
+    "characters-valid": "Seuls les lettres, les chiffres et les caractères suivants sont acceptés\u00a0: apostrophe ('), virgule (,), point (.), trait d'union (-) et parenthèse ()",
+    "city-required": "Entrez une ville ou une municipalité pour votre adresse postale",
+    "country-required": "Sélectionnez un pays pour votre adresse postale",
+    "invalid-postal-zip-code-for-country": "Un code postal canadien a été entré dans l'adresse postale mais le Canada n'a pas été choisi comme pays. Veuillez choisir Canada, et la bonne province, ou changer le code postal",
+    "invalid-postal-zip-code-for-province": "Le code postal ne correspond pas à la province ou au territoire que vous avez sélectionné",
+    "postal-zip-code-required": "Entrez un code postal ou un code postal américain pour votre adresse postale",
+    "postal-zip-code-valid": "Entrez un code postal pour l'adresse postale dans le bon format, par exemple A1A 1A1",
+    "province-state-required": "Sélectionnez une province, un territoire, un état ou une région pour votre adresse postale",
+    "zip-code-valid": "Entrez un code postal américain pour l'adresse postale dans le bon format, par exemple 12345 ou 12345-6789"
+  }
+}
